JSObjConverter call JSON.stringify in js context
Starting with Rhino 1.7.14, NativeJSON.stringify will no longer be able
to be called with a null context. By calling in javascript code, we can
also avoid needing to have a custom NativeDate class.  Furthermore,
changing canConvert to test if the type implements Scriptable removes
the need for a custom NativeBoolean and also catches classes like
NativeString and NativeNumber that were previously not accounted for.